video-processing-platform/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.27.2
pytest==9.0.2
motor==3.6.1
pymongo==4.9.2
google-generativeai==0.3.0
